live view sync changes from miro board
