+++ src/lib/ui/ZoneSelector.svelte
@@ -37,15 +37,19 @@
 </style>
 
 <div class="w-full py-1 px-2 bg-gray-100 text-gray-900 text-xl flex items-center">
+    {#if zones.length>1}
     <div class="flex justify-center arrow" class:disabled={zones.indexOf(zone)<=0} 
     on:click={()=>{if(!(zones.indexOf(zone)<=0)) {zone=zones[zones.indexOf(zone)-1]}}}>
         <span class="text-xl mx-2">&lt;</span>
     </div>
+    {/if}
     <div class="flex justify-center text-ellipsis grow">
         {zone}
     </div>
+    {#if zones.length>1}
     <div class="flex justify-center arrow" class:disabled={zones.indexOf(zone)+1>=zones.length} 
         on:click={()=>{if(!(zones.indexOf(zone)+1>=zones.length)) {zone=zones[zones.indexOf(zone)+1]}}}>
         <span class="text-xl mx-2">&gt;</span>
     </div>
+    {/if}
 </div>
